LIBSEARCH-39. Database entries without a hostUrl get a hyperlink of "#"
Some Database Finder entries point to services that are no
no longer available, but are kept for historical/informational
purposes, such as "LexisNexis Academic".

Since the NCSU QuickSearch code wants to suppress results without
hostUrls, this modification sets a default hyperlink of "#" for
entries without a hostUrl. Clicking the title link will send the
user to the top of the page.

https://issues.umd.edu/browse/LIBSEARCH-39